refactored events into dedicated Events class
